orchestrator/requirements.txt
anthropic>=0.40.0
openai>=1.50.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
psutil>=6.0.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
